Code review feedback. Additional validation cases. Simplify StartupPropertyHandler and direct subclasses by moving common impl into base class.
